// src/pages/modules/website-utama/publication/List-data/data/types.ts
export interface IPublication {
  id_publikasi: string
  id_tahun_publikasi: string
  id_satuan_organisasi: string

  nama_publikasi: string
  slug: string
  penulis: string
  link: string

  created_at: string
  created_user: string
  updated_at: string
  updated_user: string

  nama_user_created: string
  nama_user_updated: string

  nama_tahun_publikasi: string
}